update temp drawer component name
