+++ docs/src/components/componentList/drawerComponent/index.tsx
@@ -1,3 +1,3 @@
 export * from './DrawerDocsComponent';
 export * from './DrawerPlaygrounPage';
-export * from './DrawerComponentPlayground';
+export * from './DrawerComponentPreview';
